+++ packages/lightspark-sdk/src/objects/LightningTransaction.ts
@@ -98,6 +98,7 @@ export const LightningTransactionFromJson = (
       obj["outgoing_payment_uma_post_transaction_data"]?.map((e) =>
         PostTransactionDataFromJson(e),
       ),
+      obj["outgoing_payment_payment_preimage"],
     );
   }
   if (obj["__typename"] == "RoutingTransaction") {
@@ -460,6 +461,7 @@ fragment LightningTransactionFragment on LightningTransaction {
                 currency_amount_preferred_currency_value_approx: preferred_currency_value_approx
             }
         }
+        outgoing_payment_payment_preimage: payment_preimage
     }
     ... on RoutingTransaction {
         __typename