+++ packages/lightspark-sdk/src/objects/ScreenNodeInput.ts
@@ -3,8 +3,13 @@
 import ComplianceProvider from "./ComplianceProvider.js";
 
 type ScreenNodeInput = {
+  /**
+   * The compliance provider that is going to screen the node. You need to be a customer of the selected
+   * provider and store the API key on the Lightspark account setting page.
+   **/
   provider: ComplianceProvider;
 
+  /** The public key of the lightning node that needs to be screened. **/
   nodePubkey: string;
 };