+++ packages/ui/styles/fonts/typography/Code.tsx
@@ -1,30 +1,21 @@
 "use client";
 
 import styled from "@emotion/styled";
-import { App, getTypographyString, TokenSize } from "../typographyTokens";
+import { getTypographyString, TokenSize } from "../typographyTokens";
 
 interface Props {
   children: React.ReactNode;
-  app?: App;
   size?: TokenSize;
 }
 
-export const Code = ({
-  children,
-  app = App.Lightspark,
-  size = TokenSize.Medium,
-}: Props) => {
-  return (
-    <CodeStyles app={app} size={size}>
-      {children}
-    </CodeStyles>
-  );
+export const Code = ({ children, size = TokenSize.Medium }: Props) => {
+  return <CodeStyles size={size}>{children}</CodeStyles>;
 };
 
 const CodeStyles = styled.div<Props>`
-  ${({ theme, app, size }) => {
-    return app && size
-      ? getTypographyString(theme.typography[app].Code[size])
+  ${({ theme, size }) => {
+    return size
+      ? getTypographyString(theme.typography[theme.app].Code[size])
       : "";
   }}
 `;